services/user-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
firebase-admin==6.2.0
pydantic==2.4.2
python-multipart==0.0.6
google-cloud-firestore==2.13.0
google-cloud-secret-manager==2.16.3
python-dotenv==1.0.0
# Testing dependencies
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.1 